Add multer for download images

+++ backend/controllers/stuff.js
@@ -1,10 +1,17 @@
 // controllers/stuff.js
 const Thing = require("../models/Thing");
+const fs = require("fs");
 
 exports.createThing = (req, res, next) => {
-  delete req.body._id;
+  const thingObject = JSON.parse(req.body.thing);
+  delete thingObject._id;
+  delete thingObject._userId;
   const thing = new Thing({
-    ...req.body,
+    ...thingObject,
+    userId: req.body.userId,
+    imageUrl: `${req.protocol}://${req.get("host")}/images/${
+      req.file.filename
+    }`,
   });
   thing
     .save()
@@ -19,19 +26,48 @@ exports.getOneThing = (req, res, next) => {
 };
 
 exports.modifyThing = (req, res, next) => {
-  Thing.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() =>
-      res.status(200).json({ message: "Object modified successfully!" })
-    )
-    .catch((error) => res.status(400).json({ error }));
+  const thingObject = reqfile
+    ? {
+        ...JSON.parse(req.body.thing),
+        imageUrl: `${req.protocol}://${req.get("host")}/images/${
+          req.file.filename
+        }`,
+      }
+    : { ...req.body };
+  delete thingObject._userId;
+  Thing.findOne({ _id: req.params.id })
+    .then((thing) => {
+      if (thing.userId !== req.body.userId) {
+        return res.status(401).json({ error: "Unauthorized" });
+      } else {
+        Thing.updateOne(
+          { _id: req.params.id },
+          { ...thingObject, _id: req.params.id }
+        )
+          .then(() =>
+            res.status(200).json({ message: "Object modified successfully!" })
+          )
+          .catch((error) => res.status(400).json({ error }));
+      }
+    })
+    .catch((error) => res.status(500).json({ error }));
 };
 
 exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({ _id: req.params.id })
-    .then(() =>
-      res.status(200).json({ message: "Object deleted successfully!" })
-    )
-    .catch((error) => res.status(400).json({ error }));
+  Thing.findOne({ _id: req.params.id }).then((thing) => {
+    if (thing.userId !== req.auth.userId) {
+      return res.status(401).json({ error: "Unauthorized" });
+    } else {
+      const filename = thing.imageUrl.split("/images/")[1];
+      fs.unlink(`images/${filename}`, () => {
+        Thing.deleteOne({ _id: req.params.id })
+          .then(() =>
+            res.status(200).json({ message: "Object deleted successfully!" })
+          )
+          .catch((error) => res.status(400).json({ error }));
+      });
+    }
+  });
 };
 
 exports.getAllThings = (req, res, next) => {
